perf(materiales): las páginas del pager dejan de remontarse en cada render
IntroPageItem y ContentPageItem se definían dentro del render de
MaterialPagesScreen: cada render del padre creaba un TIPO de componente
nuevo, así que React desmontaba y remontaba cada página visible del pager
en vez de actualizarla (scroll perdido, círculos decorativos
re-aleatorizados).

Hoisted a components/materiales/MaterialPageItems.tsx con props explícitas
(el módulo-level in-file superaba las 400 líneas del propio archivo, así
que se extrae a carpeta por la cláusula del plan). renderItem sin cambios
de comportamiento, solo pasa las props nuevas.

Plan 009 de plans/README.md — marcado DONE. Sin entrada de CHANGELOG
(refactor interno sin cambio funcional).
