+++ mcm-app/app/screens/MaterialPagesScreen.tsx
@@ -3,12 +3,8 @@ import {
   View,
   StyleSheet,
   Dimensions,
-  Text,
   FlatList,
-  ScrollView,
   Platform,
-  DimensionValue,
-  ViewStyle,
   type ColorSchemeName,
 } from 'react-native';
 import { RouteProp } from 'expo-router/react-navigation';
@@ -16,16 +12,19 @@ import colors, { Colors } from '@/constants/colors';
 import { useColorScheme } from '@/hooks/useColorScheme';
 import useFontScale from '@/hooks/useFontScale';
 import spacing from '@/constants/spacing';
-import FormattedContent from '@/components/FormattedContent';
+import {
+  IntroPageItem,
+  ContentPageItem,
+} from '@/components/materiales/MaterialPageItems';
 import { MasStackParamList } from '../(tabs)/mas';
 
-interface Pagina {
+export interface Pagina {
   titulo?: string;
   subtitulo?: string;
   texto?: string;
 }
 
-interface Actividad {
+export interface Actividad {
   id: string;
   nombre: string;
   emoji: string;
@@ -34,26 +33,7 @@ interface Actividad {
 }
 
 type RouteProps = RouteProp<MasStackParamList, 'MaterialPages'>;
-
-const generateRandomCircles = (count: number = 5) => {
-  const circles = [];
-  const SIZES = [50, 100, 150, 80, 120, 60, 200, 400]; // Variety of sizes
-
-  for (let i = 0; i < count; i++) {
-    const size = SIZES[i % SIZES.length];
-    const opacity = Math.random() * (0.15 - 0.05) + 0.05; // Random opacity between 0.05 and 0.15
-    const top = `${Math.random() * 120 - 10}%`; // Random top between -10% and 110%
-    const left = `${Math.random() * 120 - 10}%`; // Random left between -10% and 110%
-    circles.push({
-      size,
-      opacity,
-      top,
-      left,
-      color: '#FFF',
-    });
-  }
-  return circles;
-};
+export type MaterialPagesStyles = ReturnType<typeof createStyles>;
 
 export default function MaterialPagesScreen({ route }: { route: RouteProps }) {
   const scrollTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null);
@@ -67,125 +47,31 @@ export default function MaterialPagesScreen({ route }: { route: RouteProps }) {
     [scheme, introBackgroundColor, fontScale],
   );
 
-  const IntroPageItem = ({ actividad }: { actividad: Actividad }) => {
-    const circlesData = React.useMemo(() => generateRandomCircles(5), []);
-    return (
-      <View style={[styles.introPage, { width }]}>
-        {circlesData.map((circle, idx) => (
-          <View
-            key={`deco-${idx}`}
-            style={
-              {
-                position: 'absolute',
-                width: circle.size,
-                height: circle.size,
-                borderRadius: circle.size / 2,
-                backgroundColor: circle.color,
-                opacity: circle.opacity,
-                top: circle.top as DimensionValue,
-                left: circle.left as DimensionValue,
-              } as ViewStyle
-            }
-          />
-        ))}
-        <Text style={styles.introEmoji} selectable>
-          {actividad.emoji}
-        </Text>
-        <Text style={styles.introTitle} selectable>
-          {actividad.nombre.toUpperCase()}
-        </Text>
-        <Text style={styles.introDate} selectable>
-          {new Date(fecha)
-            .toLocaleDateString('es-ES', {
-              weekday: 'long',
-              day: 'numeric',
-              month: 'long',
-            })
-            .replace(',', '')
-            .toUpperCase()}
-        </Text>
-        <Text style={styles.introHint} selectable>
-          Desliza para ver el material
-        </Text>
-      </View>
-    );
-  };
-
-  const ContentPageItem = ({ item }: { item: Pagina }) => {
-    const content = item.texto ? (
-      <FormattedContent text={item.texto} scale={fontScale} />
-    ) : null;
-
-    if (Platform.OS === 'web') {
-      // Calculate the remaining height after header, dots, and tab bar
-      const headerHeight = 80; // Approximate header height
-      const dotsHeight = 60; // Approximate dots container height
-      const tabBarHeight = 80; // Approximate tab bar height at bottom
-      const contentHeight = height - headerHeight - dotsHeight - tabBarHeight;
-
-      return (
-        <View style={[styles.page, { width, height }]}>
-          <View
-            style={[styles.pageHeader, { backgroundColor: actividad.color }]}
-          >
-            <Text style={styles.pageTitle} selectable>
-              {item.titulo}
-            </Text>
-            {item.subtitulo && (
-              <Text style={styles.pageSubtitle} selectable>
-                {item.subtitulo}
-              </Text>
-            )}
-          </View>
-          <div
-            style={{
-              height: contentHeight,
-              overflowY: 'auto',
-              padding: spacing.lg,
-              boxSizing: 'border-box',
-              marginBottom: dotsHeight, // Add margin to avoid overlap with dots
-            }}
-          >
-            {content}
-          </div>
-        </View>
-      );
-    }
-
-    return (
-      <View style={[styles.page, { width }]}>
-        <View style={[styles.pageHeader, { backgroundColor: actividad.color }]}>
-          <Text style={styles.pageTitle} selectable>
-            {item.titulo}
-          </Text>
-          {item.subtitulo && (
-            <Text style={styles.pageSubtitle} selectable>
-              {item.subtitulo}
-            </Text>
-          )}
-        </View>
-        <ScrollView
-          contentContainerStyle={styles.pageContent}
-          showsVerticalScrollIndicator={true}
-          bounces={true}
-          overScrollMode={'auto'}
-          style={{ flex: 1 }}
-          scrollEnabled={true}
-        >
-          {content}
-        </ScrollView>
-      </View>
-    );
-  };
   const [index, setIndex] = useState(0);
   const pages = [{ intro: true }, ...actividad.paginas];
   const { width, height } = Dimensions.get('window');
 
   const renderItem = ({ item }: { item: any }) => {
     if (item.intro) {
-      return <IntroPageItem actividad={actividad} />;
+      return (
+        <IntroPageItem
+          actividad={actividad}
+          styles={styles}
+          width={width}
+          fecha={fecha}
+        />
+      );
     }
-    return <ContentPageItem item={item as Pagina} />;
+    return (
+      <ContentPageItem
+        item={item as Pagina}
+        styles={styles}
+        fontScale={fontScale}
+        width={width}
+        height={height}
+        color={actividad.color}
+      />
+    );
   };
 
   const getItemLayout = (_data: any, itemIndex: number) => ({
